Redimension, suppression images et best books
- Sharp + multer pour récup et redimensionner les images depuis Files
- Suppression des images lors de la suppression du livre
- Suppression des anciennes images lors de la modification de celles-ci

+++ backend/controllers/books.js
@@ -1,15 +1,31 @@
 const Book = require("../models/book");
+const sharp = require("sharp");
+const fs = require("fs");
+
+const resizeImage = (filepath, filename) => {
+  sharp(filepath)
+    .resize(405, 570)
+    .toFormat("webp")
+    .webp({ quality: 80 })
+    .toFile("./images/resized/" + filename);
+};
 
 exports.addBook = (req, res, next) => {
   const bookObject = JSON.parse(req.body.book);
+  let ref = "";
+  if (req.file) {
+    ref = req.file.filename + ".webp";
+    const path = req.file.path;
+    resizeImage(path, ref);
+  } else {
+    return res.status(400).json({ message: "Fichier image manquant !" });
+  }
   delete bookObject._id;
   delete bookObject._userId;
   const book = new Book({
     ...bookObject,
     userId: req.auth.userId,
-    imageUrl: `${req.protocol}://${req.get("host")}/images/${
-      req.file.filename
-    }`,
+    imageUrl: `${req.protocol}://${req.get("host")}/images/resized/${ref}`,
   });
   book
     .save()
@@ -22,12 +38,16 @@ exports.addRating = (req, res, next) => {
 };
 
 exports.modifyBook = (req, res, next) => {
+  let ref = "";
+  if (req.file) {
+    ref = req.file.filename + ".webp";
+    const path = req.file.path;
+    resizeImage(path, ref);
+  }
   const bookObject = req.file
     ? {
         ...JSON.parse(req.body.book),
-        imageUrl: `${req.protocol}://${req.get("host")}/images/${
-          req.file.filename
-        }`,
+        imageUrl: `${req.protocol}://${req.get("host")}/images/resized/${ref}`,
       }
     : { ...req.body };
   delete bookObject._userId;
@@ -36,6 +56,13 @@ exports.modifyBook = (req, res, next) => {
       if (book.userId !== req.auth.userId) {
         res.status(401).json({ message: "Non-autorisé" });
       } else {
+        if (req.file) {
+          const resizedFilename = book.imageUrl.split("/images/resized/")[1];
+          const originalFilename = resizedFilename.split(".webp")[0];
+          fs.unlink(`images/${originalFilename}`, () => {
+            fs.unlinkSync(`images/resized/${resizedFilename}`);
+          });
+        }
         Book.updateOne(
           { _id: req.params.id },
           { ...bookObject, _id: req.params.id }
@@ -48,9 +75,23 @@ exports.modifyBook = (req, res, next) => {
 };
 
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Livre supprimé !" }))
-    .catch((error) => res.status(400).json({ error }));
+  Book.findOne({ _id: req.params.id })
+    .then((book) => {
+      if (book.userId !== req.auth.userId) {
+        res.status(401).json({ message: "Not authorized" });
+      } else {
+        const resizedFilename = book.imageUrl.split("/images/resized/")[1];
+        const originalFilename = resizedFilename.split(".webp")[0];
+        fs.unlink(`images/${originalFilename}`, () => {
+          fs.unlink(`images/resized/${resizedFilename}`, () => {
+            Book.deleteOne({ _id: req.params.id })
+              .then(() => res.status(200).json({ message: "Livre supprimé !" }))
+              .catch((error) => res.status(401).json({ error }));
+          });
+        });
+      }
+    })
+    .catch((error) => res.status(500).json({ error }));
 };
 
 exports.getOneBook = (req, res, next) => {
@@ -60,7 +101,15 @@ exports.getOneBook = (req, res, next) => {
 };
 
 exports.getBestBooks = (req, res, next) => {
-  // recupère les 3 livres les mieux notés (moyenne)
+  Book.find()
+    .then((books) => {
+      books.sort((a, b) => {
+        return a.averageRating - b.averageRating;
+      });
+      const bestBooks = books.slice(-3);
+      res.status(200).json(bestBooks);
+    })
+    .catch((error) => res.status(400).json({ error }));
 };
 
 exports.getAllBooks = (req, res, next) => {
